refactor: separated the tests of each core functionality into separate files
