+++ test/core/black-white-filter.test.js
@@ -1,0 +1,30 @@
+import convertToBW from "../../src/core/black-white-filter.js"
+
+describe('convertToBW function', () => {
+    test('converts pixel array to black and white correctly', () => {
+        const pixelArray = [255, 255, 255, 255, 0, 0, 0, 255]; // RGBA values
+        const expectedResult = [255, 255, 255, 255, 0, 0, 0, 255]; // Expected black and white pixel array
+        expect(convertToBW(pixelArray)).toEqual(expectedResult);
+    });
+
+    test('handles empty input array', () => {
+        const pixelArray = [];
+        expect(convertToBW(pixelArray)).toEqual([]);
+    });
+
+    test('handles random RGBA values correctly', () => {
+        // Generate a random pixel array with RGBA values between 0 and 255
+        const randomPixelArray = Array.from({ length: 20 }, () => Math.floor(Math.random() * 256));
+        const pixelArray = [...randomPixelArray.slice(0, 3), 255, ...randomPixelArray.slice(4)]; // Ensure alpha channel is 255
+        const bwPixelArray = convertToBW(pixelArray);
+
+        // Ensure that each RGB value in the converted array is either 0 or 255
+        const isValidBWPixelArray = bwPixelArray.every((value, index) => {
+            // Skip alpha channel (every fourth value)
+            if ((index + 1) % 4 === 0) return true;
+            return value === 0 || value === 255;
+        });
+
+        expect(isValidBWPixelArray).toBe(true);
+    });
+});
