+++ test/core/gray-scale-filter.test.js
@@ -1,0 +1,29 @@
+import convertToGrayScale from "../../src/core/gray-scale-filter.js";
+
+describe('convertToGrayScale function', () => {
+    test('converts pixel array to grayscale correctly', () => {
+        const pixelArray = [255, 255, 255, 255, 0, 0, 0, 255]; // RGBA values
+        const expectedGrayScaleArray = [255, 255, 255, 255, 0, 0, 0, 255]; // Expected grayscale pixel array
+        expect(convertToGrayScale(pixelArray)).toEqual(expectedGrayScaleArray);
+    });
+
+    test('handles empty input array', () => {
+        const pixelArray = [];
+        expect(convertToGrayScale(pixelArray)).toEqual([]);
+    });
+
+    test('handles random RGBA values correctly', () => {
+        // Generate a random pixel array with RGBA values between 0 and 255
+        const randomPixelArray = Array.from({ length: 20 }, () => Math.floor(Math.random() * 256));
+        const pixelArray = [...randomPixelArray.slice(0, 3), 255, ...randomPixelArray.slice(4)]; // Ensure alpha channel is 255
+        const grayScalePixelArray = convertToGrayScale(pixelArray);
+
+        // Ensure that each set of four consecutive values represents a valid grayscale color
+        for (let i = 0; i < grayScalePixelArray.length; i += 4) {
+            const grayValue = grayScalePixelArray[i];
+            expect(grayScalePixelArray[i]).toBe(grayValue); // R component
+            expect(grayScalePixelArray[i + 1]).toBe(grayValue); // G component
+            expect(grayScalePixelArray[i + 2]).toBe(grayValue); // B component
+        }
+    });
+});
